pyproject.toml
[tool.poetry]
name = "school-practice-nicola"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
mysql-connector-python = "^8.0.25"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
